Создал получение отп кода, обработку во viewModel. А также реализовал хранилище для токена, стейт аутентификации. Возможно в будущем перенесу хранилище с shared preferences на data store

+++ app/src/main/java/com/example/pizza_shift_2025/authorization_screen/domain/model/OtpResponse.kt
@@ -3,5 +3,5 @@ package com.example.pizza_shift_2025.authorization_screen.domain.model
 class OtpResponse(
     val success: Boolean,
     val errorReason: String,
-    val retryDelayInSeconds: Long
+    val retryDelayInSeconds: Int
 )
